requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
asyncpg==0.29.0
bcrypt==4.2.0
click==8.1.7
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pylint==3.2.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==13.0
